Refactor: 덕포인트 패칭 위치 변경
