+++ src/containers/main/UserProfile.tsx
@@ -1,3 +1,5 @@
+'use client'
+
 import Image from 'next/image'
 import { useSession } from 'next-auth/react'
 import img from '@/dummydata/profile.png'
